new: added methods for Class, fix: pass arity to Function

+++ objects/common.h
@@ -3,7 +3,7 @@
 
 typedef Value (*next_builtin_fn)(const Value *args);
 
-enum Visibility : int { PUBLIC = 0, PRIVATE = 1 };
+enum Visibility : uint8_t { PUBLIC = 0, PRIVATE = 1 };
 
 #define NEXT_BUILTIN(type, name) Value next_##type##_##name(const Value *args)
